All common actions moved from LocalStatisticService to StatisticService
